refactor: rename db folder as dynamo-db
